+++ jest.config.ts
@@ -5,6 +5,7 @@ const sharedConfig = {
         // use typescript to convert from esm to cjs
         '[.](m|c)?(ts|js)(x)?$': ['ts-jest', {
             'isolatedModules': true,
+            'tsconfig': 'tsconfig.jest.json'
         }],
     },
     // any tests that operate on dist files shouldn't compile them again.