+++ src/source/worker.ts
@@ -1,4 +1,4 @@
-import {Actor} from '../util/actor';
+import {Actor, ActorTarget} from '../util/actor';
 import {StyleLayerIndex} from '../style/style_layer_index';
 import {VectorTileWorkerSource} from './vector_tile_worker_source';
 import {TemporalGridTileWorkerSource} from './temporalgrid_tile_worker_source';
@@ -25,7 +25,7 @@ import type {PluginState} from './rtl_text_plugin';
  * The Worker class responsidble for background thread related execution
  */
 export default class Worker {
-    self: WorkerGlobalScopeInterface;
+    self: WorkerGlobalScopeInterface & ActorTarget;
     actor: Actor;
     layerIndexes: {[_: string]: StyleLayerIndex};
     availableImages: {[_: string]: Array<string>};
@@ -48,7 +48,7 @@ export default class Worker {
     };
     referrer: string;
 
-    constructor(self: WorkerGlobalScopeInterface) {
+    constructor(self: WorkerGlobalScopeInterface & ActorTarget) {
         this.self = self;
         this.actor = new Actor(self, this);
 
@@ -224,13 +224,13 @@ export default class Worker {
         return layerIndexes;
     }
 
-    getWorkerSource(mapId: string, type: string, source: string) {
+    getWorkerSource(mapId: string, sourceType: string, sourceName: string): WorkerSource {
         if (!this.workerSources[mapId])
             this.workerSources[mapId] = {};
-        if (!this.workerSources[mapId][type])
-            this.workerSources[mapId][type] = {};
+        if (!this.workerSources[mapId][sourceType])
+            this.workerSources[mapId][sourceType] = {};
 
-        if (!this.workerSources[mapId][type][source]) {
+        if (!this.workerSources[mapId][sourceType][sourceName]) {
             // use a wrapped actor so that we can attach a target mapId param
             // to any messages invoked by the WorkerSource
             const actor = {
@@ -238,10 +238,10 @@ export default class Worker {
                     this.actor.send(type, data, callback, mapId);
                 }
             };
-            this.workerSources[mapId][type][source] = new (this.workerSourceTypes[type] as any)((actor as any), this.getLayerIndex(mapId), this.getAvailableImages(mapId));
+            this.workerSources[mapId][sourceType][sourceName] = new (this.workerSourceTypes[sourceType] as any)((actor as any), this.getLayerIndex(mapId), this.getAvailableImages(mapId));
         }
 
-        return this.workerSources[mapId][type][source];
+        return this.workerSources[mapId][sourceType][sourceName];
     }
 
     getDEMWorkerSource(mapId: string, source: string) {