+++ src/style/style.test.ts
@@ -13,7 +13,7 @@ import {
 } from '../source/rtl_text_plugin';
 import {browser} from '../util/browser';
 import {OverscaledTileID} from '../source/tile_id';
-import {fakeXhr, fakeServer} from 'nise';
+import {fakeServer, type FakeServer} from 'nise';
 
 import {EvaluationParameters} from './evaluation_parameters';
 import {LayerSpecification, GeoJSONSourceSpecification, FilterSpecification, SourceSpecification} from '@globalfishingwatch/maplibre-gl-style-spec';
@@ -78,21 +78,17 @@ function createStyle(map = getStubMap()) {
     return style;
 }
 
-let sinonFakeXMLServer;
-let sinonFakeServer;
+let server: FakeServer;
 let mockConsoleError;
 
 beforeEach(() => {
     global.fetch = null;
-    sinonFakeServer = fakeServer.create();
-    sinonFakeXMLServer = fakeXhr.useFakeXMLHttpRequest();
-
+    server = fakeServer.create();
     mockConsoleError = jest.spyOn(console, 'error').mockImplementation(() => { });
 });
 
 afterEach(() => {
-    sinonFakeXMLServer.restore();
-    sinonFakeServer.restore();
+    server.restore();
     mockConsoleError.mockRestore();
 });
 
@@ -115,12 +111,12 @@ describe('Style', () => {
 
     test('loads plugin immediately if already registered', done => {
         clearRTLTextPlugin();
-        sinonFakeServer.respondWith('/plugin.js', 'doesn\'t matter');
+        server.respondWith('/plugin.js', 'doesn\'t matter');
         setRTLTextPlugin('/plugin.js', (error) => {
             expect(error).toMatch(/Cannot set the state of the rtl-text-plugin when not in the web-worker context/);
             done();
         });
-        sinonFakeServer.respond();
+        server.respond();
         new Style(getStubMap());
     });
 
@@ -152,7 +148,7 @@ describe('Style', () => {
             jest.spyOn(style.sourceCaches['vector'], 'reload');
 
             clearRTLTextPlugin();
-            sinonFakeServer.respondWith('/plugin.js', 'doesn\'t matter');
+            server.respondWith('/plugin.js', 'doesn\'t matter');
             const _broadcast = style.dispatcher.broadcast;
             style.dispatcher.broadcast = function (type, state, callback) {
                 if (type === 'syncRTLPluginState') {
@@ -171,7 +167,7 @@ describe('Style', () => {
                     done();
                 }, 0);
             });
-            sinonFakeServer.respond();
+            server.respond();
         });
     });
 });
@@ -211,15 +207,15 @@ describe('Style#loadURL', () => {
         });
 
         style.loadURL('style.json');
-        sinonFakeServer.respondWith(JSON.stringify(createStyleJSON({version: 'invalid'})));
-        sinonFakeServer.respond();
+        server.respondWith(JSON.stringify(createStyleJSON({version: 'invalid'})));
+        server.respond();
     });
 
     test('cancels pending requests if removed', () => {
         const style = new Style(getStubMap());
         style.loadURL('style.json');
         style._remove();
-        expect(sinonFakeServer.lastRequest.aborted).toBe(true);
+        expect((server.lastRequest as any).aborted).toBe(true);
     });
 });
 
@@ -266,21 +262,8 @@ describe('Style#loadJSON', () => {
         // stub Image so we can invoke 'onload'
         // https://github.com/jsdom/jsdom/commit/58a7028d0d5b6aacc5b435daee9fd8f9eacbb14c
 
-        // fake the image request (sinon doesn't allow non-string data for
-        // server.respondWith, so we do so manually)
-        const requests = [];
-        sinonFakeXMLServer.onCreate = req => { requests.push(req); };
-        const respond = () => {
-            let req = requests.find(req => req.url === 'http://example.com/sprite.png');
-            req.setStatus(200);
-            req.response = new ArrayBuffer(8);
-            req.onload();
-
-            req = requests.find(req => req.url === 'http://example.com/sprite.json');
-            req.setStatus(200);
-            req.response = '{}';
-            req.onload();
-        };
+        server.respondWith('GET', 'http://example.com/sprite.png', new ArrayBuffer(8));
+        server.respondWith('GET', 'http://example.com/sprite.json', '{}');
 
         const style = new Style(getStubMap());
 
@@ -303,7 +286,7 @@ describe('Style#loadJSON', () => {
                 done();
             });
 
-            respond();
+            server.respond();
         });
     });
 
@@ -315,21 +298,8 @@ describe('Style#loadJSON', () => {
         // stub Image so we can invoke 'onload'
         // https://github.com/jsdom/jsdom/commit/58a7028d0d5b6aacc5b435daee9fd8f9eacbb14c
 
-        // fake the image request (sinon doesn't allow non-string data for
-        // server.respondWith, so we do so manually)
-        const requests = [];
-        sinonFakeXMLServer.onCreate = req => { requests.push(req); };
-        const respond = () => {
-            let req = requests.find(req => req.url === 'http://example.com/sprite.png');
-            req.setStatus(200);
-            req.response = new ArrayBuffer(8);
-            req.onload();
-
-            req = requests.find(req => req.url === 'http://example.com/sprite.json');
-            req.setStatus(200);
-            req.response = '{"image1": {"width": 1, "height": 1, "x": 0, "y": 0, "pixelRatio": 1.0}}';
-            req.onload();
-        };
+        server.respondWith('GET', 'http://example.com/sprite.png', new ArrayBuffer(8));
+        server.respondWith('GET', 'http://example.com/sprite.json', '{"image1": {"width": 1, "height": 1, "x": 0, "y": 0, "pixelRatio": 1.0}}');
 
         const style = new Style(getStubMap());
 
@@ -357,7 +327,7 @@ describe('Style#loadJSON', () => {
                 });
             });
 
-            respond();
+            server.respond();
         });
     });
 
@@ -751,7 +721,7 @@ describe('Style#setState', () => {
     });
 
     test('Issue #3893: compare new source options against originally provided options rather than normalized properties', done => {
-        sinonFakeServer.respondWith('/tilejson.json', JSON.stringify({
+        server.respondWith('/tilejson.json', JSON.stringify({
             tiles: ['http://tiles.server']
         }));
         const initial = createStyleJSON();
@@ -767,7 +737,7 @@ describe('Style#setState', () => {
             style.setState(initial);
             done();
         });
-        sinonFakeServer.respond();
+        server.respond();
     });
 
     test('return true if there is a change', done => {