+++ src/style/style_glyph.ts
@@ -1,14 +1,15 @@
 import type {AlphaImage} from '../util/image';
 
-/**
- * The glyph's metrices
- */
 export type GlyphMetrics = {
     width: number;
     height: number;
     left: number;
     top: number;
     advance: number;
+    /**
+     * isDoubleResolution = true for 48px textures
+     */
+    isDoubleResolution?: boolean;
 };
 
 /**