+++ test/integration/browser/browser.test.ts
@@ -2,20 +2,21 @@ import puppeteer, {Page, Browser} from 'puppeteer';
 import st from 'st';
 import http from 'http';
 import type {Server} from 'http';
-import fs from 'fs';
-import path from 'path';
-import pixelmatch from 'pixelmatch';
-import {PNG} from 'pngjs';
 import type {AddressInfo} from 'net';
+import type {Map} from '../../../src/ui/map';
+import type {default as MapLibreGL} from '../../../src/index';
 
 const testWidth = 800;
 const testHeight = 600;
+const deviceScaleFactor = 2;
 
 let server: Server;
 let browser: Browser;
 let page: Page;
-let map: any;
-let maplibregl: any;
+let map: Map;
+let maplibregl: typeof MapLibreGL;
+
+jest.retryTimes(3);
 
 describe('Browser tests', () => {
 
@@ -32,7 +33,7 @@ describe('Browser tests', () => {
 
     beforeEach(async () => {
         page = await browser.newPage();
-        await page.setViewport({width: testWidth, height: testHeight, deviceScaleFactor: 2});
+        await page.setViewport({width: testWidth, height: testHeight, deviceScaleFactor});
 
         const port = (server.address() as AddressInfo).port;
 
@@ -47,6 +48,17 @@ describe('Browser tests', () => {
                 }
             });
         });
+    }, 40000);
+
+    afterEach(async() => {
+        page.close();
+    }, 40000);
+
+    afterAll(async () => {
+        await browser.close();
+        if (server) {
+            server.close();
+        }
     }, 40000);
 
     test('Load should fire before resize and moveend', async () => {
@@ -87,23 +99,36 @@ describe('Browser tests', () => {
     }, 20000);
 
     test('Drag to the left', async () => {
-
         const canvas = await page.$('.maplibregl-canvas');
         const canvasBB = await canvas?.boundingBox();
 
-        // Perform drag action, wait a bit the end to avoid the momentum mode.
-        await page.mouse.move(canvasBB!.x, canvasBB!.y);
-        await page.mouse.down();
-        await page.mouse.move(100, 0);
-        await new Promise(r => setTimeout(r, 200));
-        await page.mouse.up();
+        const dragToLeft = async () => {
+            await page.mouse.move(canvasBB!.x, canvasBB!.y);
+            await page.mouse.down();
+            await page.mouse.move(100, 0, {
+                steps: 10
+            });
+            await page.mouse.up();
+            await new Promise(r => setTimeout(r, 200));
 
-        const center = await page.evaluate(() => {
-            return map.getCenter();
-        });
+            return page.evaluate(() => {
+                return map.getCenter();
+            });
+        };
 
-        expect(center.lng).toBeCloseTo(-35.15625, 4);
-        expect(center.lat).toBeCloseTo(0, 7);
+        await page.emulateMediaFeatures([
+            {name: 'prefers-reduced-motion', value: 'reduce'},
+        ]);
+        const centerWithoutInertia = await dragToLeft();
+        expect(centerWithoutInertia.lng).toBeCloseTo(-35.15625, 4);
+        expect(centerWithoutInertia.lat).toBeCloseTo(0, 7);
+
+        await page.emulateMediaFeatures([
+            {name: 'prefers-reduced-motion', value: 'reduce'},
+        ]);
+        const centerWithInertia = await dragToLeft();
+        expect(centerWithInertia.lng).toBeLessThan(-60);
+        expect(centerWithInertia.lat).toBeCloseTo(0, 7);
     }, 20000);
 
     test('Resize viewport (page)', async () => {
@@ -148,97 +173,9 @@ describe('Browser tests', () => {
         expect(zoom).toBe(2);
     }, 20000);
 
-    test('CJK Characters', async () => {
-
-        await page.evaluate(() => {
-
-            map.setStyle({
-                version: 8,
-                glyphs: 'https://mierune.github.io/fonts/{fontstack}/{range}.pbf',
-                sources: {
-                    sample: {
-                        type: 'geojson',
-                        data: {
-                            type: 'Feature',
-                            geometry: {
-                                type: 'Point',
-                                coordinates: [0, 0]
-                            },
-                            properties: {
-                                'name_en': 'abcde',
-                                'name_ja': 'あいうえお',
-                                'name_ch': '阿衣乌唉哦',
-                                'name_kr': '아이우'
-                            }
-                        }
-                    },
-                },
-                'layers': [
-                    {
-                        'id': 'sample-text-left',
-                        'type': 'symbol',
-                        'source': 'sample',
-                        'layout': {
-                            'text-anchor': 'top',
-                            'text-field': '{name_ja}{name_en}',
-                            'text-font': ['Open Sans Regular'],
-                            'text-offset': [-10, 0],
-                        }
-                    },
-                    {
-                        'id': 'sample-text-center',
-                        'type': 'symbol',
-                        'source': 'sample',
-                        'layout': {
-                            'text-anchor': 'top',
-                            'text-field': '{name_ch}{name_kr}',
-                            'text-font': ['Open Sans Regular'],
-                            'text-offset': [0, 0],
-                        }
-                    },
-                    {
-                        'id': 'sample-text-right',
-                        'type': 'symbol',
-                        'source': 'sample',
-                        'layout': {
-                            'text-anchor': 'top',
-                            'text-field': '{name_en}{name_ja}',
-                            'text-font': ['Open Sans Regular'],
-                            'text-offset': [10, 0],
-                        }
-                    },
-                ]
-            });
-        });
-
-        const image = await page.evaluate(() => {
-            return new Promise((resolve, _) => {
-                map.once('idle', () => resolve(map.getCanvas().toDataURL()));
-                map.setZoom(8);
-            });
-        });
-
-        const actualBuff = Buffer.from((image as string).replace(/data:.*;base64,/, ''), 'base64');
-        const actualPng = new PNG({width: testWidth, height: testHeight});
-        actualPng.parse(actualBuff);
-
-        const expectedPlatforms = ['ubuntu-runner', 'macos-runner', 'macos-local'];
-        let minDiff = Infinity;
-        for (const expected of expectedPlatforms) {
-            const diff = compareByPixelmatch(actualPng, expected, testWidth, testHeight);
-            if (diff < minDiff) {
-                minDiff = diff;
-            }
-        }
-
-        // At least one platform should be identical
-        expect(minDiff).toBe(0);
-
-    }, 20000);
-
     test('Marker: correct position', async () => {
         const markerScreenPosition = await page.evaluate(() => {
-            const markerMapPosition = [11.40, 47.30];
+            const markerMapPosition = [11.40, 47.30] as [number, number];
             const marker = new maplibregl.Marker()
                 .setLngLat(markerMapPosition)
                 .addTo(map);
@@ -311,35 +248,4 @@ describe('Browser tests', () => {
         expect(markerScreenPosition.x).toBeCloseTo(386.5);
         expect(markerScreenPosition.y).toBeCloseTo(378.1);
     }, 20000);
-
-    afterEach(async() => {
-        page.close();
-    }, 40000);
-
-    afterAll(async () => {
-        await browser.close();
-        if (server) {
-            server.close();
-        }
-    }, 40000);
-
-    function compareByPixelmatch(actualPng:PNG, platform: string, width:number, height:number): number {
-        const platformFixtureBase64 = fs.readFileSync(
-            path.join(__dirname, `fixtures/cjk-expected-base64-image/${platform}-base64.txt`), 'utf8')
-            .replace(/\s/g, '')
-            .replace(/data:.*;base64,/, '');
-
-        const expectedBuff = Buffer.from(platformFixtureBase64, 'base64');
-
-        const expectedPng = new PNG({width: testWidth, height: testHeight});
-        expectedPng.parse(expectedBuff);
-
-        const diffImg = new PNG({width, height});
-
-        const diff = pixelmatch(
-            actualPng.data, expectedPng.data, diffImg.data,
-            width, height, {threshold: 0}) / (width * height);
-
-        return diff;
-    }
 });